--- a/paddle/fluid/operators/activation_op.cu
+++ b/paddle/fluid/operators/activation_op.cu
@@ -18,9 +18,7 @@ namespace paddle {
 namespace operators {
 
 using Tensor = framework::Tensor;
-// result = a > 0 ? b : 0;
-template <typename T>
-__device__ __forceinline__ T reluFunc(const T* a, const T* b);
+using float16 = paddle::platform::float16;
 
 template <typename T, int N>
 struct GetVecType;
@@ -29,33 +27,54 @@ template <typename T>
 struct GetVecType<T, 1> {
   using type = T;
 };
+
 template <>
 struct GetVecType<platform::float16, 2> {
-  using type = half2;
+  using type = __half2;
 };
 
 template <>
 struct GetVecType<float, 4> {
   using type = float4;
 };
-// for relu forward and backward
-template <>
-__device__ __forceinline__ double reluFunc<double>(const double* a,
-                                                   const double* b) {
-  return ((*a) > 0.0f) * (*b);
-}
+
+template <typename T>
+class BaseActivationGPUFunctor {
+ public:
+  using ELEMENT_TYPE = T;
+};
+
+template <typename Type>
+class ReluGPUFunctor : public BaseActivationGPUFunctor<Type> {
+ public:
+  // for relu forward and backward when T is double
+  template <typename T>
+  __device__ __forceinline__ T compute(const T* a, const T* b) {
+#ifdef __CUDA_ARCH__ >= 350 || HIP_VERSION >= 300
+    return __ldg(a) > static_cast<T>(0) ? __ldg(b) : static_cast<T>(0);
+#else
+    return (*a) > static_cast<T>(0) ? (*b) : static_cast<T>(0);
+#endif
+  }
+
+  // when num % vecsize != 0 this func will be used
+  template <typename T>
+  __device__ __forceinline__ T computeReminder(const T a, const T b) {
+    return a > static_cast<T>(0) ? b : static_cast<T>(0);
+  }
+};
 
 template <>
-__device__ __forceinline__ float4 reluFunc<float4>(const float4* a,
-                                                   const float4* b) {
+__device__ __forceinline__ float4
+ReluGPUFunctor<float>::compute<float4>(const float4* a, const float4* b) {
   return make_float4((a->x > 0.0f) * (b->x), (a->y > 0.0f) * (b->y),
                      (a->z > 0.0f) * (b->z), (a->w > 0.0f) * (b->w));
 }
 
 template <>
-__device__ __forceinline__ __half2 reluFunc<__half2>(const __half2* a,
-                                                     const __half2* b) {
-#if __CUDA_ARCH__ >= 530 || CUDA_VERSION >= 300
+__device__ __forceinline__ __half2
+ReluGPUFunctor<float16>::compute<__half2>(const __half2* a, const __half2* b) {
+#ifdef __CUDA_ARCH__ >= 530 || CUDA_VERSION >= 300
   const half2 kzero = __float2half2_rn(0.0f);
   return __hmul2(__hgt2(__ldg(a), kzero), __ldg(b));
 #else
@@ -66,114 +85,30 @@ __device__ __forceinline__ __half2 reluFunc<__half2>(const __half2* a,
 #endif
 }
 
-// relu forward kernel
-template <typename T, typename VecType, int VECSIZE>
-__global__ void reluKernelCudaVec(const T* in, T* out, int num) {
+template <typename T, typename VecType, int VecSize, typename Functor>
+__global__ void reluKernelCudaVec(const T* src1, const T* src2, T* dst, int num,
+                                  Functor functor) {
   int idx = threadIdx.x + blockIdx.x * blockDim.x;
   int stride = blockDim.x * gridDim.x;
-  int loop = num / VECSIZE;
-  int tail = num % VECSIZE;
-  const VecType* src = reinterpret_cast<const VecType*>(in);
-  VecType* dst = reinterpret_cast<VecType*>(out);
+  int loop = num / VecSize;
+  int tail = num % VecSize;
+  const VecType* in1 = reinterpret_cast<const VecType*>(src1);
+  const VecType* in2 = reinterpret_cast<const VecType*>(src2);
+  VecType* out = reinterpret_cast<VecType*>(dst);
+
   for (int i = idx; i < loop; i += stride) {
-    dst[i] = reluFunc((src + i), (src + i));
+    out[i] = functor.compute((in1 + i), (in2 + i));
   }
-  T temp;
-  T zero = (T)(0.0f);
+
   while (idx == loop && tail) {
-    temp = in[num - tail];
-    out[num - tail] = temp > zero ? temp : zero;
+    dst[num - tail] =
+        functor.computeReminder(src1[num - tail], src2[num - tail]);
     --tail;
   }
 }
-// relu backward kernel
-template <typename T, typename VecType, int VECSIZE>
-__global__ void reluGradKernelVec(const T* out, const T* dout, T* dx, int num) {
-  int idx = threadIdx.x + blockIdx.x * blockDim.x;
-  int stride = blockDim.x * gridDim.x;
-  int loop = num / VECSIZE;
-  int tail = num % VECSIZE;
-  const VecType* out_v = reinterpret_cast<const VecType*>(out);
-  const VecType* dout_v = reinterpret_cast<const VecType*>(dout);
-  VecType* dx_out = reinterpret_cast<VecType*>(dx);
-  for (int i = idx; i < loop; i += stride) {
-    dx_out[i] = reluFunc((out_v + i), (dout_v + i));
-  }
-  T zero = (T)(0.0f);
-  while (idx == loop && tail) {
-    dx[num - tail] = out[num - tail] > zero ? dout[num - tail] : zero;
-    --tail;
-  }
-}
-template <typename T>
-__global__ void reluGradFunc(const T* out, const T* dout, T* dx, int num) {
-  const float4* out_f = reinterpret_cast<const float4*>(out);
-  const float4* dout_f = reinterpret_cast<const float4*>(dout);
-  float4* dx_f = reinterpret_cast<float4*>(dx);
-  int i = blockDim.x * blockIdx.x + threadIdx.x;
-  int loop = num / 4;
-  int tail = num % 4;
-  for (int t = i; t < loop; t += blockDim.x * gridDim.x) {
-    dx_f[t].x = (out_f[t].x > 0.0f) * dout_f[t].x;
-    dx_f[t].y = (out_f[t].y > 0.0f) * dout_f[t].y;
-    dx_f[t].z = (out_f[t].z > 0.0f) * dout_f[t].z;
-    dx_f[t].w = (out_f[t].w > 0.0f) * dout_f[t].w;
-  }
-  T zero = (T)(0.0f);
-  while (i == loop && tail) {
-    dx[num - tail] = out[num - tail] > zero ? dout[num - tail] : zero;
-    --tail;
-  }
-}
-template <typename T>
-__global__ void reluGradFunc_half(const T* out, const T* dout, T* dx,
-                                  size_t num) {
-  int idx = threadIdx.x + blockIdx.x * blockDim.x;
-  if (idx * 2 >= num) return;
-  int loop = num >> 1;
-  int stride = blockDim.x * gridDim.x;
-  const __half2* out_half = reinterpret_cast<const __half2*>(out);
-  const __half2* dout_half = reinterpret_cast<const __half2*>(dout);
-  __half2* dx_half = reinterpret_cast<__half2*>(dx);
 
-  const half2 kzero = __float2half2_rn(0.0f);
-
-  for (int i = idx; i < loop; i += stride) {
-#if __CUDA_ARCH__ >= 530 || CUDA_VERSION >= 300
-    dx_half[i] =
-        __hmul2(__hgt2(__ldg(out_half + i), kzero), __ldg(dout_half + i));
-#else
-    const float2 xx = __half22float2(out[i]);
-    const float2 yy = __half22float2(dout[i]);
-    dx_half[i] =
-        __floats2half2_rn(xx.x > 0.0f ? static_cast<float>(yy.x) : 0.0f,
-                          xx.y > 0.0f ? static_cast<float>(yy.y) : 0.0f);
-#endif
-  }
-}
-
-template <typename T, int VECSIZE>
-struct ReluGPUFunctor : public BaseActivationFunctor<T> {
-  void operator()(const platform::CUDADeviceContext& context,
-                  const framework::Tensor& in, framework::Tensor* out,
-                  int num) {
-    const T* input_data = in.data<T>();
-    T* output_data = out->mutable_data<T>(context.GetPlace());
-    using VecType = typename GetVecType<T, VECSIZE>::type;
-    int block = 512;
-    if (num > VECSIZE) {
-      int grid = (num / VECSIZE + block - 1) / block;
-      reluKernelCudaVec<T, VecType, VECSIZE><<<grid, block>>>(input_data,
-                                                              output_data, num);
-    } else {
-      int grid1 = (num + block - 1) / block;
-      reluKernelCudaVec<T, T, 1><<<grid1, block>>>(input_data, output_data,
-                                                   num);
-    }
-  }
-};
-template <typename DeviceContext, typename Functor>
-class ReluBaseKernel
+template <typename DeviceContext, typename Functor, int VecSize>
+class ActivationGPUKernel
     : public framework::OpKernel<typename Functor::ELEMENT_TYPE> {
  public:
   using T = typename Functor::ELEMENT_TYPE;
@@ -181,39 +116,28 @@ class ReluBaseKernel
     const Tensor* in_x = context.Input<Tensor>("X");
     Tensor* out = context.Output<Tensor>("Out");
     auto& dev_ctx = context.template device_context<DeviceContext>();
+
     int num = in_x->numel();
-    Functor functor;
-    functor(dev_ctx, *in_x, out, num);
-  }
-};
+    const T* input_data = in_x->data<T>();
+    T* output_data = out->mutable_data<T>(dev_ctx.GetPlace(),
+                                          static_cast<size_t>(num * sizeof(T)));
 
-template <typename T, int VECSIZE>
-struct ReluGradGPUFunctor : public BaseActivationFunctor<T> {
-  void operator()(const platform::CUDADeviceContext& context,
-                  const framework::Tensor* out, const framework::Tensor* d_out,
-                  framework::Tensor* d_x) {
-    auto numel = d_out->numel();
-    auto* dout_data = d_out->data<T>();
-    auto* out_data = out->data<T>();
-    auto* dx_data = d_x->mutable_data<T>(
-        context.GetPlace(), static_cast<size_t>(numel * sizeof(T)));
-    using VecType = typename GetVecType<T, VECSIZE>::type;
     int block = 512;
-    if (numel > VECSIZE) {
-      int grid = (numel / VECSIZE + block - 1) / block;
-      reluGradKernelVec<T, VecType, VECSIZE><<<grid, block>>>(
-          out_data, dout_data, dx_data, numel);
-    } else {
-      int grid1 = (numel + block - 1) / block;
-      reluGradKernelVec<T, T, 1><<<grid1, block>>>(out_data, dout_data, dx_data,
-                                                   numel);
+    int grid = (num + block - 1) / block;
+    if (num > VecSize) {
+      // update grid for float16 and float32
+      grid = (num / VecSize + block - 1) / block;
     }
+
+    Functor functor;
+    using VecType = typename GetVecType<T, VecSize>::type;
+    reluKernelCudaVec<T, VecType, VecSize, Functor><<<grid, block>>>(
+        input_data, input_data, output_data, num, functor);
   }
-  // static constexpr ActBwdOpFwdDeps FwdDeps() { return kDepOut; }
 };
 
-template <typename DeviceContext, typename Functor>
-class ReluGradKernel
+template <typename DeviceContext, typename Functor, int VecSize>
+class ActivationGradGPUKernel
     : public framework::OpKernel<typename Functor::ELEMENT_TYPE> {
  public:
   using T = typename Functor::ELEMENT_TYPE;
@@ -223,9 +147,25 @@ class ReluGradKernel
         ctx.Input<framework::Tensor>(framework::GradVarName("Out"));
     framework::Tensor* d_x =
         ctx.Output<framework::Tensor>(framework::GradVarName("X"));
-    Functor functor;
+
     auto& dev_ctx = ctx.template device_context<DeviceContext>();
-    functor(dev_ctx, out, d_out, d_x);
+    auto* out_data = out->data<T>();
+    auto numel = d_out->numel();
+    auto* dout_data = d_out->data<T>();
+    auto* dx_data = d_x->mutable_data<T>(
+        dev_ctx.GetPlace(), static_cast<size_t>(numel * sizeof(T)));
+
+    int block = 512;
+    int grid = (numel + block - 1) / block;
+    if (numel >= VecSize) {
+      // update grid for float16 and float32
+      grid = (numel / VecSize + block - 1) / block;
+    }
+
+    Functor functor;
+    using VecType = typename GetVecType<T, VecSize>::type;
+    reluKernelCudaVec<T, VecType, VecSize, Functor><<<grid, block>>>(
+        out_data, dout_data, dx_data, numel, functor);
   }
 };
 
@@ -234,7 +174,6 @@ class ReluGradKernel
 
 namespace ops = paddle::operators;
 namespace plat = paddle::platform;
-
 #define REGISTER_ACTIVATION_CUDA_KERNEL(act_type, op_name, functor,         \
                                         grad_functor)                       \
   REGISTER_OP_CUDA_KERNEL(                                                  \
@@ -281,20 +220,21 @@ REGISTER_OP_CUDA_KERNEL(
 
 /* ===========================    relu register  ============================ */
 REGISTER_OP_CUDA_KERNEL(
-    relu, ops::ReluBaseKernel<paddle::platform::CUDADeviceContext,
-                              ops::ReluGPUFunctor<float, 4>>,
-    ops::ReluBaseKernel<paddle::platform::CUDADeviceContext,
-                        ops::ReluGPUFunctor<double, 1>>,
-    ops::ReluBaseKernel<paddle::platform::CUDADeviceContext,
-                        ops::ReluGPUFunctor<plat::float16, 2>>);
+    relu, ops::ActivationGPUKernel<paddle::platform::CUDADeviceContext,
+                                   ops::ReluGPUFunctor<float>, 4>,
+    ops::ActivationGPUKernel<paddle::platform::CUDADeviceContext,
+                             ops::ReluGPUFunctor<double>, 1>,
+    ops::ActivationGPUKernel<plat::CUDADeviceContext,
+                             ops::ReluGPUFunctor<plat::float16>, 2>);
 
 REGISTER_OP_CUDA_KERNEL(
-    relu_grad, ops::ReluGradKernel<plat::CUDADeviceContext,
-                                   ops::ReluGradGPUFunctor<float, 4>>,
-    ops::ReluGradKernel<plat::CUDADeviceContext,
-                        ops::ReluGradGPUFunctor<double, 1>>,
-    ops::ReluGradKernel<plat::CUDADeviceContext,
-                        ops::ReluGradGPUFunctor<plat::float16, 2>>);
+    relu_grad, ops::ActivationGradGPUKernel<paddle::platform::CUDADeviceContext,
+                                            ops::ReluGPUFunctor<float>, 4>,
+    ops::ActivationGradGPUKernel<paddle::platform::CUDADeviceContext,
+                                 ops::ReluGPUFunctor<double>, 1>,
+    ops::ActivationGradGPUKernel<plat::CUDADeviceContext,
+                                 ops::ReluGPUFunctor<plat::float16>, 2>);
+
 REGISTER_OP_CUDA_KERNEL(
     relu_grad_grad,
     ops::ActivationDoubleGradKernel<paddle::platform::CUDADeviceContext,